+++ src/main/java/fr/siamois/repositories/RecordingUnitStudyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecordingUnitStudyRepository extends CrudRepository<RecordingUnitStudy, Integer> {
+public interface RecordingUnitStudyRepository extends CrudRepository<RecordingUnitStudy, Long> {
 
 
 }
