+++ src/main/java/fr/siamois/repositories/VocabularyTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VocabularyTypeRepository extends CrudRepository<VocabularyType, Integer> {
+public interface VocabularyTypeRepository extends CrudRepository<VocabularyType, Long> {
 
 
 }
